feat: Backend integration & Type safety improvements
## Backend (New)
- FastAPI backend with Python 3.12
- Yahoo Finance integration (yahooquery with Ghost Mode)
- TEFAS fund data integration
- APScheduler for periodic data fetching
- Professional type stubs for third-party libraries

## Flutter Integration
- BackendApiService for API communication
- Platform-aware URLs (Android emulator support)
- Watchlist with SharedPreferences persistence
- 30-second auto-refresh for market data
- Swipe-to-delete functionality

## Code Quality
- Type annotations throughout backend
- Modern FastAPI lifespan pattern
- Pylance configuration for clean IDE
- Custom type stubs for yahooquery, apscheduler, tefas

## Bug Fixes
- Fixed USD/TRY, EUR/TRY parsing (pair/rate fields)
- Fixed delete button instant refresh
- Fixed chart selector 2x2 grid layout

+++ lib/widgets/loading_widgets.dart
@@ -61,7 +61,7 @@ class _SkeletonLoaderState extends State<SkeletonLoader>
                   _controller.value - 0.3,
                   _controller.value,
                   _controller.value + 0.3,
-                ].map((e) => e.clamp(0.0, 1.0) as double).toList(),
+                ].map((e) => e.clamp(0.0, 1.0)).toList(),
               ).createShader(bounds);
             },
             blendMode: BlendMode.srcATop,
